refactor(ingredients): Remove sub-ingredients feature to fix app instability
This commit completely removes the sub-ingredients feature from the application. This feature was identified by the user as the source of a critical bug that caused a `ReferenceError` in the frontend production build and led to application instability.

The following changes were made to fully revert the feature:
- Removed the `subIngredients` property from the Zod schema in `shared/schema.ts`.
- Removed all logic for processing and displaying sub-ingredients from the utility functions in `client/src/lib/ingredient-utils.ts`.
- Removed all UI elements, state, and handler functions related to managing sub-ingredients from the `ingredients-step.tsx` component.

This resolves the reported error by reverting the application to a stable state before the problematic feature was introduced.
